requirements.txt
# ForensiX AI Requirements

# Core

numpy>=1.21.0
python-magic>=0.4.24

# Image Analysis

Pillow>=9.0.0

# Network Analysis

scapy>=2.4.5

# Machine Learning (Optional - للتطوير المستقبلي)

scikit-learn>=1.0.0
tensorflow>=2.8.0

# NLP (Optional)

nltk>=3.6

# Report Generation

reportlab>=3.6.0
markdown>=3.3.0
